refactor(dash-network): move `FFINetwork` + extern fn into the new crate
The C-ABI mirror of `Network` belongs with its owning type. Moves
`dash/src/ffi/network.rs` verbatim (plus additional unit tests) into
`dash-network/src/ffi.rs` behind a new `ffi` feature flag. `dashcore`'s
existing `ffi` feature now forwards to `dash-network/ffi`, and
`dash/src/ffi/mod.rs` re-exports so the stable public path
`dashcore::ffi::FFINetwork` / `dashcore::ffi::dashcore_network_get_name`
is preserved for all downstream consumers (dash-spv-ffi, etc.).

Added tests in the new location cover:
- ABI discriminants (Mainnet=0, Testnet=1, Devnet=2, Regtest=3)
- `FFINetwork <-> Network` round-trip
- `dashcore_network_get_name` returns canonical lowercase names

The extern symbol `dashcore_network_get_name` is still exported via
the dashcore rlib (verified with `nm`).
