+++ dash-network/Cargo.toml
@@ -18,6 +18,8 @@ default = []
 serde = ["dep:serde"]
 # bincode::{Encode, Decode} impls for Network.
 bincode = ["dep:bincode", "dep:bincode_derive"]
+# C-ABI `FFINetwork` mirror + `dashcore_network_get_name` extern fn.
+ffi = []
 
 [dependencies]
 serde = { version = "1.0.219", default-features = false, features = ["derive"], optional = true }
