+++ dash/Cargo.toml
@@ -33,7 +33,7 @@ eddsa = ["ed25519-dalek"]
 quorum_validation = ["bls"]
 message_verification = ["bls"]
 bincode = [ "dep:bincode", "dep:bincode_derive", "dashcore_hashes/bincode", "dash-network/bincode" ]
-ffi = []
+ffi = ["dash-network/ffi"]
 test-utils = []
 
 [package.metadata.docs.rs]
