+++ dash/src/ffi/mod.rs
@@ -1,3 +1,6 @@
-mod network;
+//! C-compatible FFI surface of `dashcore`.
+//!
+//! `FFINetwork` now lives in the `dash-network` crate; it is re-exported here
+//! so that `dashcore::ffi::FFINetwork` remains the stable public path.
 
-pub use network::FFINetwork;
+pub use dash_network::ffi::{FFINetwork, dashcore_network_get_name};
